docker_node_installer_remote.py changes: - Line 42: Set a new variable for the ip address of the dtr.  This is a temp fix just to test.  When it works, I will test the dtrHost to see if it is an ip address. - Line 99: Set dtrIp equal to host (which is an ip) - Line 166: Setting the hostname equal to the value gotten from the dtrIp - Lines 169-170: Added conditional to set the dtrHost to the hostname if the dtrHost and dtrIp match (this means they would both be ip's)
dtr_installer_remote.py changes:
- Lines 33-34: Put in conditional to set dtrHost equal to the hostname
if the dtrHost and dtrIp match (this means they would both be ip's)
- Line 36: Set --ucp-node to hostname.  As it turns out this variable
must always be the host name and not the domain name, b/c this is how
the node is registered in the ucp.  The external url is where the domain
name is set.